requirements.txt
tensorflow>=2.8.0
scikit-learn>=1.0.0
imbalanced-learn>=0.8.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
numpy>=1.21.0
Pillow>=8.0.0
tqdm>=4.62.0
streamlit>=1.10.0